Input File
This is the raw input file. The column we care about is column AV. It is the products details section. In a perfect world, every portion of that product details section would be its own column (like product name, product weight etc.) but if we can do it just for primary reason for purchase and how you heard about us that would be amazing. And then I can manipulate the program as need be to get to the other stuff.